Fixed a bug in BlameTreeShadowObject (reference to field not declared).
Removed templates. In general, I don't think it is a good idea to use templates, plus now I don't see the need to have to specify different precision for the array that holds results in different precisions. We can certainly put this back if there is a strong preference and/or need for it. 

-- line, and those below, will be ignored--

M    BlameTreeShadowObject.h
M    BlameTree.cpp
M    BlameTree.h


git-svn-id: https://repo.eecs.berkeley.edu/svn/users/ksen/projects/corvette-instrumentor@6435 1c3fd406-9068-4342-8353-511fcdf43622